requirements.txt
scikit-learn==0.24.2
matplotlib==3.3.3
psycopg2-binary==2.9.9
sendgrid==6.10.0
pandas==2.0.2
secure-smtplib==0.1.1
tensorflow==2.9.2
apache-airflow-providers-apache-spark==4.1.5
pyspark==3.5.0